requirements.txt
Django==1.10.1
django-cors-headers==1.2.1
djangorestframework==3.4.7
djangorestframework-jwt==1.8.0
djoser==0.5.1
mysqlclient==1.3.9
PyJWT==1.4.2
